Added gravitational force calcule

+++ src/Class.hpp
@@ -3,17 +3,17 @@
 class Sun
 {
 public:
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
     Sun();
 };
 
 class Earth
 {
 public:
-    float x; 
-    float y;
-    float z;
+    double x; 
+    double y;
+    double z;
     Earth();
 };
